Fix bugs in chunked_even
Previous `maxbuf` was too small for `n < 3` and larger than necessary for `n > 3`. `n + (n-2)(n-1)` is necessary and sufficient.

Sufficient: consider, without loss of generality, that there are `r in [0, n - 1]` elements left in the iterable (if `r >= n`, yield lists of size `n` until `r < n`). If we yield a list of `n` now, are we guaranteed to be able to yield all lists of either `n` or `n-1` from now on, for any possible `r`? Start by yielding a list, which leaves `(n-2)(n-1)` in the buffer, i.e. `(n-2)` lists of `(n-1)` elements. Then, if we were to read in those `r` elements:
(a) If `r = n - 1`, we would have `(n-1)(n-1)`, i.e. `(n-1)` lists of `(n-1)` elements.
(b) If `r <= n - 2`, we would have `r` lists of `n` elements, and `((n-2) - r)` lists of `(n-1)` elements.

Necessary: if we use a smaller buffer, `n + (n-2)(n-1) - 1`, and yield a list, `(n-2)(n-1) - 1 = X` elements remain. Consider the case that there are no more elements (`r = 0`). Is it possible to split `X` into lists of size `n` and `n-1`, i.e. are there `A >= 0, B >= 0` s.t. `X = A n + B (n-1)`? This is a Diophantine equation. Since `gcd(n, n-1) = 1`, it has solutions. Specifically, if `1 = A' n + B' (n-1)`, then `A = X A'` and `B = X B'`. `1 = A' n + B' (n-1)` has a solution `A' = 1, B' = -1`, so in general solutions look like `A' = 1 + k n, B' = -1 - k (n - 1)`. Therefore, either `A` or `B` is negative, so this smaller buffer size is not enough.